automation-tests/requirements.txt
# AI Generated Code by Deloitte + Cursor (BEGIN)
# Selenium E2E Testing Dependencies
selenium>=4.15.0
pytest>=7.4.0
pytest-html>=4.1.0
pytest-selenium>=4.0.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
allure-pytest>=2.13.0
# AI Generated Code by Deloitte + Cursor (END)
